Add files via upload
This file trains the ANN with a sine wave, then tests the ANN with a larger amplitude, lower frequency sine wave to demonstrate its ability to keep the core temperature near 37 *C.  The ANN is trained only with the q and surface temp values after the system gets to 37 degrees.  These same q values are then used for the testing.