(1) Took out all the deprecated stuff (and there was a LOT)
(2) Added the ability to use multiple time steps when training.  Set the value Nt to anything greater than one. This required some reprogramming of the array logic because now the length of all the arrays are not the same.

(3) Changed the name of greensFromSKL method

(4) Changed the PID parameters to be a function of the Fourier number.

(5) Cleaned up some comments.